fix: use PathBuf push method to add files to the given folder
The method set_file_name swaps the last component of the file path with the given file name, even if that last component is a directory name. This was not the intended behaviour and this is fixed by using push instead, which works because we already confirm earlier in the program the given path is a directory.